submit exer 4.15 - 4.23 && authorization and token authentication

+++ part4/blog-list/controllers/blogs.js
@@ -1,25 +1,75 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/Blog');
+const jwt = require('jsonwebtoken');
+const config = require('../utils/config');
+const { userExtractor } = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user', { blogs: 0 });
   response.json(blogs);
 });
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   if(!request.body.title|| !request.body.url){
     return response.status(400).end();
   }
 
-  const blog = new Blog(request.body);
-  const result = await blog.save();
-  response.status(201).json(result);
+  const decodedToken = jwt.verify(request.token, config.SECRET);
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'invalid token' });
+  }
+
+  const user = request.user;
+
+  if(!user){
+    return response.status(401).json({ error: 'user not found ' });
+  }
+
+  const blog = new Blog({
+    ...request.body,
+    user: user._id
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndRemove(request.params.id);
-  console.log('result', result);
-  result === null ? response.status(404).end() : response.status(204).end();
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  if(!request.token){
+    return response.status(401).json({ error: 'need a token to delete ' });
+  }
+
+  const decodedToken = jwt.verify(request.token, config.SECRET);
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'invalid token' });
+  }
+
+  const user = request.user;
+
+  if(!user){
+    return response.status(401).json({ error: 'user not found ' });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  console.log('blog', blog);
+
+  if(!blog){
+    return response.status(404).end();
+  }
+
+  console.log('blog user id === user.id', user._id.toString() === blog.user.toString());
+
+  if(! (user && user._id.toString() === blog.user.toString())){
+    return response.status(401).json({ error: 'user unaouzaried to delete this blog' });
+  }
+  console.log('blog', blog);
+  await Blog.findByIdAndRemove(request.params.id);
+
+  response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -37,5 +87,6 @@ blogsRouter.put('/:id', async (request, response) => {
   console.log('updatedBlog', updatedBlog);
   response.json(updatedBlog);
 });
+
 
 module.exports = blogsRouter;
